requirements_minimal.txt
# Minimal requirements for quick local setup
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Optional document processing (install as needed)
PyPDF2==3.0.1
python-docx==0.8.11

# For enhanced features (optional)
requests==2.31.0
python-dotenv==1.0.0
